// generated/src/app/api-models/sminstallation-s.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SminstallationS {
  namedobjectIdSequenceid: number;
  namedobjectIdObjecttype: number;
  installationOraclehome: string;
  installationLanguage: string;
  installationOs: string;
  installationOsversion: string;
  installationMachinetype: number;
}